feat: move template to ts
